Handle YYFILL labels with --loop-switch option.
In order to avoid splitting YYFILL states in two and adding more jumps
between them, use --eager-skip option. This option moves skip statements
from the beginning of each state to transitions into this state from
other states. This means that YYFILL label can now be attached to the
beginning of the state (no need to bypass the skip statement), and there
is no need to split states.

The code generated with -f, --storable-state option can be optimized to
use one state variable, and the initial switch on states can be simpler.